+++ changelog.md
@@ -4,6 +4,7 @@ Saves from 8.x are not compatible with 9.0.0.
 
 ## Features/Improvements
 
+* **[Flight Planning]** Improved IP selection for targets that are near the center of a threat zone.
 * **[Modding]** Factions can now specify the ship type to be used for cargo shipping. The Handy Wind will be used by default, but WW2 factions can pick something more appropriate.
 
 ## Fixes